requirements.txt
newsapi-python==0.2.7
vaderSentiment==3.3.2
python-dotenv==1.0.0
rich==13.7.0
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
google-generativeai==0.3.2 